ENH: Build ITK with ITKVtkGlue support.
The ITKVtkGlue ITK module adds filters that can be inserted into pipelines
that convert itk::Image to vtkImageData and vtkImageData to itk::Image. This
is very useful for using VTK and ITK code together.  These filters can be
inserted into a pipeline, which prevents the need to write an entire class
that wraps functionality for every filter/pipeline in Libs/vtkITK.