COMP: Add VTK6 find_package COMPONENT calls.
The ITKVtkGlue module call find_package(VTK with COMPONENTS, so each VTK
sub-project needs to call find_package(VTK with COMPONENTS that use different
components. This change is necessary to avoid linking errors.

The components are complicated, because they change based on the OpenGL
implementation requested, etc. For now, use the inclusive set of Slicer
COMPONENTS defined at the top level. This can be paired down in the future.