ENH: Add TravisCI builds of pull requests.
When a pull request comes in on GitHub, this tests that Slicer builds without
errors. To save time, it only builds Slicer and Slicer's dependencies built
with the superbuild.  Slicer's depedencies are pre-built in a Docker image.
This patch does not run the tests because of a lack of a graphical
environment. If the branch contains changes to the Superbuild/ directory, the
build intentionally always fails to indicate that it does not yet test these
changes.